Convert to JSON; refactor tests; re-arrange modules
