+++ data/sources/pluto.json
@@ -1,10 +1,11 @@
 {
-  "id": "zoning-districts",
+  "id": "pluto",
   "type": "cartovector",
+  "minzoom": 12,
   "source-layers": [
     {
-      "id": "zoning-districts",
-      "sql": "SELECT * FROM (\n              SELECT *, CASE\n                WHEN SUBSTRING(zonedist, 3, 1) = '-' THEN LEFT(zonedist, 2)\n                WHEN SUBSTRING(zonedist, 3, 1) ~ E'[A-Z]' THEN LEFT(zonedist, 2)\n                WHEN SUBSTRING(zonedist, 3, 1) ~ E'[0-9]' THEN LEFT(zonedist, 3)\n                ELSE zonedist\n              END as primaryzone FROM zoning_districts_v201806\n            ) a"
+      "id": "pluto",
+      "sql": "SELECT the_geom_webmercator, bbl, lot, landuse, address FROM mappluto_v18_1"
     }
   ]
 }
