+++ data/sources/zoning-districts.json
@@ -1,11 +1,10 @@
 {
-  "id": "pluto",
+  "id": "zoning-districts",
   "type": "cartovector",
-  "minzoom": 12,
   "source-layers": [
     {
-      "id": "pluto",
-      "sql": "SELECT the_geom_webmercator, bbl, lot, landuse, address FROM mappluto_v18_1"
+      "id": "zoning-districts",
+      "sql": "SELECT * FROM (\n              SELECT *, CASE\n                WHEN SUBSTRING(zonedist, 3, 1) = '-' THEN LEFT(zonedist, 2)\n                WHEN SUBSTRING(zonedist, 3, 1) ~ E'[A-Z]' THEN LEFT(zonedist, 2)\n                WHEN SUBSTRING(zonedist, 3, 1) ~ E'[0-9]' THEN LEFT(zonedist, 3)\n                ELSE zonedist\n              END as primaryzone FROM zoning_districts_v201806\n            ) a"
     }
   ]
 }
