+++ utils/structure-carto-source.js
@@ -1,16 +1,13 @@
 const carto = require('./carto');
 
-const getSource = sourceId => new Promise(async (resolve, reject) => {
-  // import sourceConfig
-  const sourceConfig = require(`../data/sources/${sourceId}.js`); // eslint-disable-line
-
+const getSource = sourceConfig => new Promise(async (resolve, reject) => {
   try {
     // instantiate carto vector tiles, get back the tile template
     const template = await carto.getVectorTileTemplate(sourceConfig['source-layers']);
 
     // make an object with sourceid as a key, and valid mapboxGL source as value
     const source = {};
-    source[sourceId] = {
+    source[sourceConfig.id] = {
       type: 'vector',
       tiles: [template],
     };
